+++ frontend/src/store/store.js
@@ -1,0 +1,8 @@
+import { configureStore } from '@reduxjs/toolkit';
+import pokerReducer from './pokerSlice';
+
+export default configureStore({
+  reducer: {
+    poker: pokerReducer,
+  },
+});