+++ CHANGELOG.md
@@ -1,4 +1,5 @@
 # Changelog
+
 All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
@@ -8,15 +9,19 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - Add React static Home and About pages, with navigation and routing
 - Add theme and styling
-- Add React Testing Library and Web Test Runner. Remove Jest.
+- Add React Testing Library and Web Test Runner and remove Jest
+- Add Redux Toolkit and React Redux
+- Add initial poker game front-end logic and store
 
 ## [0.0.1] - 2021-10-03
+
 ### Added
+
 - React site skeleton
 - Development environment based on Snowpack
 - Jest test runner
 - Hosted on Cloudflare Pages
 - Changelog
 
-[Unreleased]: https://github.com/briangershon/planning-poker/compare/v0.0.1...HEAD
+[unreleased]: https://github.com/briangershon/planning-poker/compare/v0.0.1...HEAD
 [0.0.1]: https://github.com/briangershon/planning-poker/releases/tag/v0.0.1